refactored the searchApi to take in a parameter, searchTerm, and use that in the term parameter for the /search
demonstrated manually calling the searchApi with an initial term but this causes an infinite loop as the search causes a state change which triggers a render, which calls another manual searchApi with our start search, change state, render, etc...infinite loop. Leaving the bad code in for now as history.