make use of the FlatList react-native primitive.
horizontal tells the FlatList to render horizontally instead of vertically - our spec wants scrollable to the right for the elements in tha returned array.

data is needed for FlatList - pass it the results passed in from the SearchScreen.

keyExtractor is a function to find the unique identifier of each result in the results array. We will use id from the payload.

renderItem is the function that will take in item - which is representing one item in the results array. Note that the item passed in has to be called item - it is part of the FlatList requirement. I tried to use result (single item in results array) and use result.name for the name of the restaurant but that did not work.