using the id passed in from the navigation parameter, we wil call the yelp.get with that appended.
setResult using our useState, which will cause a render. We will ONLY call getResult the first time the ResultsShowScreen is loaded, not each. This is due to the useEffect with empty array as second parameters. This one time execution will call getResult with the passed in id.

we confirmed we are getting 1 object back now from the yelp.get call.